Extracting results-converter.ts
